Split environment and viewport configuration files
This adds changes so that viewport and environment properties
are loaded from independent files.

The idea is that we don't have to repeat environment properties
when we need to change some viewport configurations.

We also now can read any property from either of the files and
not bound to only a defined set of methods.